gatekeeper/requirements.txt
fastapi 
uvicorn 
requests 
python-jose[cryptography] 
httpx
python-dotenv
psycopg2
supabase